% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_feat}
\alias{extract_feat}
\title{XML helper function:}
\usage{
extract_feat(xml, tag)
}
\arguments{
\item{xml}{An xml object}

\item{tag}{<string> Denoting the XML element to be extracted.}
}
\description{
retrieve the layer information for a WMS GetCapabilities response content.
}
